refactor: 마이페이지 사용자 조회 방식 정리
